+++ src/main/java/com/awbd/bookshop/services/bookbasket/IBookBasketService.java
@@ -3,7 +3,7 @@ package com.awbd.bookshop.services.bookbasket;
 import com.awbd.bookshop.models.Basket;
 
 public interface IBookBasketService {
-    Integer addBookToBasket(Integer bookId, Basket basket);
-    Integer removeBookToBasket(int bookId, int basketId);
-    Integer decrementBookFromBasket(int bookId, int basketId);
+    Double addBookToBasket(Integer bookId, Basket basket);
+    Double removeBookToBasket(int bookId, int basketId);
+    Double decrementBookFromBasket(int bookId, int basketId);
 }
